+++ frontend/src/pages/Canvas.tsx
