+++ frontend/src/pages/GameEnd.tsx
