+++ frontend/src/pages/GameStart.tsx
