+++ frontend/src/pages/index.ts
@@ -1,18 +1,18 @@
-export { default as Game } from './game/Game';
+export { default as Game } from './Game';
 export { default as Home } from './Home';
 export { default as NotFound } from './NotFound';
 export { default as Loading } from './Loading';
 export { default as Landing } from './Landing';
-export { default as Registration } from './auth/Registration';
-export { default as Login } from './auth/Login';
+export { default as Registration } from './Registration';
+export { default as Login } from './Login';
 export { default as ProtectedRoute } from './ProtectedRoute';
 export { default as Chat } from './Chat';
 export { default as TopRightAvatar } from './TopRightAvatar';
 export { default as PublicOnlyRoute } from './PublicOnlyRoute';
-export { default as Canvas } from './game/Canvas';
-export { default as GameEnd } from './game/GameEnd';
-export { default as GameStart } from './game/GameStart';
+export { default as Canvas } from './Canvas';
+export { default as GameEnd } from './GameEnd';
+export { default as GameStart } from './GameStart';
 export { default as Profile } from './profile/Profile';
-export { default as PrivacyPolicy } from './legal/PrivacyPolicy';
-export { default as TermsOfService } from './legal/TermsOfService';
+export { default as PrivacyPolicy } from './PrivacyPolicy';
+export { default as TermsOfService } from './TermsOfService';
 export { default as ProfileSettings } from './profile/ProfileSettings';
